requirements.txt
astroid==2.0.4
confusable-homoglyphs==3.2.0
Django==1.11
django-appconf==1.0.2
django-bootstrap4==0.0.7
django-imagekit==4.0.2
django-registration==2.4.1
django-tinymce==2.7.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pilkit==2.0
Pillow==5.3.0
psycopg2==2.7.6.1
pylint==2.1.1
python-decouple==3.1
pytz==2018.7
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
